requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
exceptiongroup==1.2.0
fake-useragent==1.5.1
greenlet==3.0.3
idna==3.7
iniconfig==2.0.0
packaging==24.0
playwright==1.43.0
pluggy==1.4.0
psycopg2-binary==2.9.9
pyee==11.1.0
pytest==8.1.1
pytest-base-url==2.1.0
pytest-playwright==0.4.4
python-dotenv==1.0.1
python-slugify==8.0.4
requests==2.31.0
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.11.0
urllib3==2.2.1
